+++ app/upgrades/v9/msg_filter_ante.go
@@ -1,4 +1,4 @@
-package ante
+package v9
 
 import (
 	"fmt"
